Fjernet kall til pensjonsinformasjonClient direkte fra BucController
flyttet litt kode fra BucController til PensjonsinformasjonService.kt
